requirements.txt
# # local package
# -e .

# external requirements
awscli
black
bs4
click
coverage
flake8
numpy
pandas
python-dotenv>=0.5.1
requests
sklearn
slack_bolt
Sphinx
wikipedia

